Fix `DelegateCodec` for recursive optional structure
In situations where we have recursive structures, we use the `DelegateCodec` to defer
codec generation. For `Optional` structures, the codec was not performing the `isPresent`
check and as a result when we tried to write the value of the empty optional, it threw
an error saying that value not found. Fixed this by overriding the `isNull` function in
`DelegateCodec` to call the codec's `isNull` function rather than running the default.